refactor(hooks): extract pure config-merge core into config_merge module
The field-level YAML merge functions — `merge_configs`, `merge_log_configs`,
`merge_hook_defs` — were buried in `yaml_config_loader`, which also owns
filesystem discovery, git-ref loading, and tracking classification. Three
consumers (the loader itself, `daft file merge`, and cross-worktree visitor
propagation) all reach for the merge semantics, so the pure core belongs in
its own module, separate from the I/O-bound loader.

Relocate the three functions verbatim into `src/hooks/config_merge.rs` (pure
— no fs, no git) along with their 8 unit tests, and split the caller imports
(`job_adapter`, `visitor_propagation`, `file/merge`, plus the loader's own
intra-module call). The no-`..` destructure guards and sparse-serialize
attributes are preserved unchanged; the function bodies are byte-identical.
